Replace input element with DatetimePicker from Nextcloud Vue
In order to add the Vue component as a region in a Marionette view a
helper class, VueWrapper, is introduced. Thanks to this a Vue instance
can be added to the current Marionette structure just by wrapping the
instance with a VueWrapper object and adding that object instead of the
Vue instance.

In turn, the Vue instance used for the lobby timer picker just contains
a DatetimePicker component from Nextcloud Vue with some appropriate
default settings (like the date format, language strings...) and a
"value" and "disabled" properties that are set and modified as needed
(for example, when the model changes or the date is being sent to the
server) from CallInfoView.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>