Work around for the datepicker pop up when its menu item is not focused
When a menu item is not focused nor hovered its opacity is reduced.
However, as the opacity of the ancestors affect the opacity of its
descendants and the datepicker pop up is a descendant of a menu item
when its menu item is not focused nor hovered the datepicker pop up has
a reduced opacity.

Moreover, as the parent pop up menu uses "filter: drop-shadow()" the
shadow also includes the shape of the datepicker pop up menu, so when
the datepicker pop up menu is not fully opaque it is darkened by the
shadow underneath.

In order to make the datepicker pop up menu fully opaque all its
ancestors must be fully opaque too, so the opacity of its menu item
needs to be forced to 1 when the pop up is shown. Unfortunately the
DatetimePicker component does not seem to emit any event to signal that
the pop up menu is shown or hidden, so the "focus" and "blur" events
emitted by the input are the closest ones available; it does not work in
every case, but it does in the most common cases.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>